requirements.txt
langgraph>=0.0.20
langchain-core>=0.1.0
ollama>=0.1.0
pymupdf>=1.23.0
pydantic>=2.0.0
typer>=0.9.0
pyyaml>=6.0
requests>=2.31.0
rich>=13.0.0

